Finish manual sending v1 tests, add unhandled error tests & fixtures
