+++ tests/fixtures/v1/manualSendNoApiKey.html
