+++ tests/fixtures/v1/unhandledError.html
@@ -1,0 +1,16 @@
+<html>
+  <head>
+    <title>Raygun4JS with V1 API</title>
+    <script src="/fixtures/common/instrumentXHRs.js"></script>
+    <script src="/dist/raygun.js"></script>
+  </head>
+  <body>
+    <script type="text/javascript">
+      Raygun.init('abcdef==').attach();
+
+      setTimeout(function () {
+        throw new Error('Unhandled error');
+      }, 1000);
+    </script>
+  </body>
+</html>
