+++ tests/specs/v1/basic.js
@@ -5,10 +5,6 @@ describe("Basic Raygun4JS V1 API tests", function() {
   // Setup
 
   beforeEach(function() {
-    //webdriverio
-    //.remote({  })
-    //.init()
-
     browser.url('http://localhost:4567/fixtures/v1/basic.html');
   });
